+++ src/comps/Home/Changelog.svelte
@@ -2,7 +2,7 @@
   import Changelog from '../../../public/CHANGELOG.md';
 </script>
 
-<details class="card small-padding">
+<details class="card small-padding" open>
   <summary><b>Changelog</b></summary>
   <Changelog />
   <a href="https://github.com/dutscher/bbx/blob/master/CHANGELOG.md" target="_blank" class="link">
@@ -15,6 +15,10 @@
     :global a,
     :global p {
       color: var(--primary);
+    }
+    :global ul {
+      margin-bottom: 8rem;
+      padding-left: 32rem;
     }
   }
 </style>